requirements.txt
a2wsgi==1.10.7
aiohappyeyeballs==2.4.4
aiohttp==3.11.8
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
asgiref==3.8.1
attrs==24.2.0
bentoml==1.3.14
cattrs==23.1.2
certifi==2024.8.30
circus==0.18.0
click==8.1.7
click-option-group==0.5.6
cloudpickle==3.1.0
deepmerge==2.0
Deprecated==1.2.15
frozenlist==1.5.0
fs==2.4.16
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
httpx-ws==0.6.2
idna==3.10
importlib-metadata==6.11.0
inflection==0.5.1
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
numpy==2.1.3
nvidia-ml-py==12.560.30
opentelemetry-api==1.20.0
opentelemetry-instrumentation==0.41b0
opentelemetry-instrumentation-aiohttp-client==0.41b0
opentelemetry-instrumentation-asgi==0.41b0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
opentelemetry-util-http==0.41b0
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
pip-requirements-parser==32.0.1
prometheus_client==0.21.0
prompt-toolkit==3.0.36
propcache==0.2.1
psutil==6.1.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-multipart==0.0.19
PyYAML==6.0.2
pyzmq==26.2.0
questionary==2.0.1
rich==13.9.4
schema==0.7.7
setuptools==75.6.0
simple-di==0.1.5
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
tomli_w==1.1.0
tornado==6.4.2
typing_extensions==4.12.2
uv==0.5.5
uvicorn==0.32.1
watchfiles==1.0.0
wcwidth==0.2.13
wrapt==1.17.0
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
